Fix permissions

+++ deployIOThermostat.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-mkdir -p /srv/http/iothermostat
+if [ $EUID -e 0 ]; then
+    echo "Please run script as regular user, not as root. Aborting."
+    exit 2
+fi
+
+sudo mkdir -p /srv/http/iothermostat
 sudo cp -R iothermostat-http/* /srv/http/iothermostat
 cp -R iothermostat-python ~
